move pipe functions to world/pipesim
